+++ astropy/coordinates/attributes.py
@@ -6,19 +6,21 @@ from __future__ import (absolute_import, unicode_literals, division,
 
 # Dependencies
 import numpy as np
+import warnings
 
 # Project
 from .. import units as u
 from ..utils.compat.numpy import broadcast_to as np_broadcast_to
+from ..utils.exceptions import AstropyDeprecationWarning
 from ..utils import OrderedDescriptor, ShapedLikeNDArray
 
-__all__ = ['FrameAttribute', 'TimeFrameAttribute', 'QuantityFrameAttribute',
+__all__ = ['Attribute', 'TimeAttribute', 'QuantityAttribute',
            'EarthLocationAttribute', 'CoordinateAttribute',
-           'CartesianRepresentationFrameAttribute',
-           'DifferentialFrameAttribute']
+           'CartesianRepresentationAttribute',
+           'DifferentialAttribute']
 
 
-class FrameAttribute(OrderedDescriptor):
+class Attribute(OrderedDescriptor):
     """A non-mutable data descriptor to hold a frame attribute.
 
     This class must be used to define frame attributes (e.g. ``equinox`` or
@@ -29,9 +31,9 @@ class FrameAttribute(OrderedDescriptor):
     The `~astropy.coordinates.FK4` class uses the following class attributes::
 
       class FK4(BaseCoordinateFrame):
-          equinox = TimeFrameAttribute(default=_EQUINOX_B1950)
-          obstime = TimeFrameAttribute(default=None,
-                                       secondary_attribute='equinox')
+          equinox = TimeAttribute(default=_EQUINOX_B1950)
+          obstime = TimeAttribute(default=None,
+                                  secondary_attribute='equinox')
 
     This means that ``equinox`` and ``obstime`` are available to be set as
     keyword arguments when creating an ``FK4`` class instance and are then
@@ -60,7 +62,7 @@ class FrameAttribute(OrderedDescriptor):
     def __init__(self, default=None, secondary_attribute=''):
         self.default = default
         self.secondary_attribute = secondary_attribute
-        super(FrameAttribute, self).__init__()
+        super(Attribute, self).__init__()
 
     def convert_input(self, value):
         """
@@ -133,10 +135,10 @@ class FrameAttribute(OrderedDescriptor):
         raise AttributeError('Cannot set frame attribute')
 
 
-class TimeFrameAttribute(FrameAttribute):
+class TimeAttribute(Attribute):
     """
     Frame attribute descriptor for quantities that are Time objects.
-    See the `~astropy.coordinates.FrameAttribute` API doc for further
+    See the `~astropy.coordinates.Attribute` API doc for further
     information.
 
     Parameters
@@ -190,7 +192,7 @@ class TimeFrameAttribute(FrameAttribute):
         return out, converted
 
 
-class CartesianRepresentationFrameAttribute(FrameAttribute):
+class CartesianRepresentationAttribute(Attribute):
     """
     A frame attribute that is a CartesianRepresentation with specified units.
 
@@ -207,7 +209,8 @@ class CartesianRepresentationFrameAttribute(FrameAttribute):
     """
 
     def __init__(self, default=None, secondary_attribute='', unit=None):
-        super(CartesianRepresentationFrameAttribute, self).__init__(default, secondary_attribute)
+        super(CartesianRepresentationAttribute, self).__init__(
+            default, secondary_attribute)
         self.unit = unit
 
     def convert_input(self, value):
@@ -244,8 +247,9 @@ class CartesianRepresentationFrameAttribute(FrameAttribute):
             # if it's a CartesianRepresentation, get the xyz Quantity
             value = getattr(value, 'xyz', value)
             if not hasattr(value, 'unit'):
-                raise TypeError('tried to set a CartesianRepresentationFrameAttribute with '
-                                'something that does not have a unit.')
+                raise TypeError('tried to set a {0} with something that does '
+                                'not have a unit.'
+                                .format(self.__class__.__name__))
 
             value = value.to(self.unit)
 
@@ -254,7 +258,7 @@ class CartesianRepresentationFrameAttribute(FrameAttribute):
             return cartrep, converted
 
 
-class QuantityFrameAttribute(FrameAttribute):
+class QuantityAttribute(Attribute):
     """
     A frame attribute that is a quantity with specified units and shape
     (optionally).
@@ -274,7 +278,7 @@ class QuantityFrameAttribute(FrameAttribute):
     """
 
     def __init__(self, default=None, secondary_attribute='', unit=None, shape=None):
-        super(QuantityFrameAttribute, self).__init__(default, secondary_attribute)
+        super(QuantityAttribute, self).__init__(default, secondary_attribute)
         self.unit = unit
         self.shape = shape
 
@@ -303,7 +307,7 @@ class QuantityFrameAttribute(FrameAttribute):
             return u.Quantity(np.zeros(self.shape), self.unit), True
         else:
             if not hasattr(value, 'unit'):
-                raise TypeError('Tried to set a QuantityFrameAttribute with '
+                raise TypeError('Tried to set a QuantityAttribute with '
                                 'something that does not have a unit.')
             oldvalue = value
             value = u.Quantity(oldvalue, self.unit, copy=False)
@@ -314,8 +318,7 @@ class QuantityFrameAttribute(FrameAttribute):
             converted = oldvalue is not value
             return value, converted
 
-
-class EarthLocationAttribute(FrameAttribute):
+class EarthLocationAttribute(Attribute):
     """
     A frame attribute that can act as a `~astropy.coordinates.EarthLocation`.
     It can be created as anything that can be transformed to the
@@ -369,7 +372,7 @@ class EarthLocationAttribute(FrameAttribute):
             return itrsobj.earth_location, True
 
 
-class CoordinateAttribute(FrameAttribute):
+class CoordinateAttribute(Attribute):
     """
     A frame attribute which is a coordinate object. It can be given as a
     low-level frame class *or* a `~astropy.coordinates.SkyCoord`, but will
@@ -426,7 +429,7 @@ class CoordinateAttribute(FrameAttribute):
             return transformedobj, True
 
 
-class DifferentialFrameAttribute(FrameAttribute):
+class DifferentialAttribute(Attribute):
     """A frame attribute which is a differential instance.
 
     The optional ``allowed_classes`` argument allows specifying a restricted
@@ -452,8 +455,8 @@ class DifferentialFrameAttribute(FrameAttribute):
         else:
             self.allowed_classes = BaseDifferential
 
-        super(DifferentialFrameAttribute, self).__init__(default,
-                                                         secondary_attribute)
+        super(DifferentialAttribute, self).__init__(default,
+                                                    secondary_attribute)
 
     def convert_input(self, value):
         """
@@ -478,13 +481,46 @@ class DifferentialFrameAttribute(FrameAttribute):
         """
 
         if not isinstance(value, self.allowed_classes):
-            raise TypeError('Tried to set a DifferentialFrameAttribute with '
+            raise TypeError('Tried to set a DifferentialAttribute with '
                             'an unsupported Differential type {0}. Allowed '
                             'classes are: {1}'
                             .format(value.__class__,
                                     self.allowed_classes))
 
         return value, True
+
+
+# Backwards-compatibility: these are the only classes that were previously
+# released in v1.3
+class FrameAttribute(Attribute):
+
+    def __init__(self, *args, **kwargs):
+        warnings.warn("FrameAttribute has been renamed to Attribute.",
+                      AstropyDeprecationWarning)
+        super(FrameAttribute, self).__init__(*args, **kwargs)
+
+class TimeFrameAttribute(TimeAttribute):
+
+    def __init__(self, *args, **kwargs):
+        warnings.warn("TimeFrameAttribute has been renamed to TimeAttribute.",
+                      AstropyDeprecationWarning)
+        super(TimeFrameAttribute, self).__init__(*args, **kwargs)
+
+class QuantityFrameAttribute(QuantityAttribute):
+
+    def __init__(self, *args, **kwargs):
+        warnings.warn("QuantityFrameAttribute has been renamed to "
+                      "QuantityAttribute.", AstropyDeprecationWarning)
+        super(QuantityFrameAttribute, self).__init__(*args, **kwargs)
+
+class CartesianRepresentationFrameAttribute(CartesianRepresentationAttribute):
+
+    def __init__(self, *args, **kwargs):
+        warnings.warn("CartesianRepresentationFrameAttribute has been renamed "
+                      "to CartesianRepresentationAttribute.",
+                      AstropyDeprecationWarning)
+        super(CartesianRepresentationFrameAttribute, self).__init__(
+            *args, **kwargs)
 
 
 # do this here to prevent a series of complicated circular imports